script: Move interfaces into smaller subfolders
All of these subfolders are tiny, which is why I bundle them
in a single commit. Otherwise the overhead of creating multiple
PRs for folders that only have 2 files is too much.

While these subfolders are tiny, the files themselves are logically
grouped together. Therefore, it's still worth it to group them,
rather than keeping them as one big blob in `script/dom`.

Part of #38901

Testing: it compiles
