+++ components/script/dom/form/mod.rs
@@ -1,0 +1,8 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod formdata;
+pub(crate) mod radionodelist;
+pub(crate) mod validation;
+pub(crate) mod validitystate;
