+++ components/script/dom/intersectionobserver/mod.rs
@@ -1,0 +1,11 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) use self::intersectionobserver::*;
+#[allow(
+    clippy::module_inception,
+    reason = "The interface name is IntersectionObserver"
+)]
+pub(crate) mod intersectionobserver;
+pub(crate) mod intersectionobserverentry;
