+++ components/script/dom/mod.rs
@@ -209,15 +209,14 @@ pub(crate) mod types {
     include!(concat!(env!("OUT_DIR"), "/InterfaceTypes.rs"));
 }
 
-pub(crate) mod abortcontroller;
-pub(crate) mod abortsignal;
+pub(crate) mod abort;
+pub(crate) use self::abort::*;
 pub(crate) mod activation;
 pub(crate) mod animations;
 pub(crate) use self::animations::*;
 pub(crate) mod audio;
 pub(crate) use self::audio::*;
 pub(crate) mod bindings;
-pub(crate) mod blob;
 #[cfg(feature = "bluetooth")]
 pub(crate) mod bluetooth;
 #[cfg(feature = "bluetooth")]
@@ -263,7 +262,8 @@ pub(crate) mod fetch;
 pub(crate) use self::fetch::*;
 pub(crate) mod file;
 pub(crate) use self::file::*;
-pub(crate) mod formdata;
+pub(crate) mod form;
+pub(crate) use self::form::*;
 pub(crate) mod fullscreen;
 #[cfg(feature = "gamepad")]
 pub(crate) mod gamepad;
@@ -280,11 +280,11 @@ pub(crate) use self::html::*;
 pub(crate) mod indexeddb;
 pub(crate) use self::indexeddb::*;
 pub(crate) mod intersectionobserver;
-pub(crate) mod intersectionobserverentry;
+pub(crate) use self::intersectionobserver::*;
 pub(crate) mod media;
 pub(crate) use self::media::*;
 pub(crate) mod mimetype;
-pub(crate) mod mimetypearray;
+pub(crate) use self::mimetype::*;
 pub(crate) mod mutationobserver;
 pub(crate) use self::mutationobserver::*;
 pub(crate) mod navigator;
@@ -295,14 +295,12 @@ pub(crate) use self::node::*;
 pub(crate) mod notification;
 pub(crate) mod performance;
 pub(crate) use self::performance::*;
-pub(crate) mod permissions;
-pub(crate) mod permissionstatus;
+pub(crate) mod permission;
+pub(crate) use self::permission::*;
 pub(crate) mod processingoptions;
-#[expect(dead_code)]
 pub(crate) mod promise;
-pub(crate) mod promisenativehandler;
+pub(crate) use self::promise::*;
 pub(crate) mod quotaexceedederror;
-pub(crate) mod radionodelist;
 pub(crate) mod range;
 pub(crate) use self::range::*;
 pub(crate) mod raredata;
@@ -322,7 +320,7 @@ pub(crate) mod servointernals;
 pub(crate) mod servoparser;
 pub(crate) mod shadowroot;
 pub(crate) mod storage;
-pub(crate) mod storagemanager;
+pub(crate) use self::storage::*;
 pub(crate) mod stream;
 pub(crate) use self::stream::*;
 pub(crate) mod svg;
@@ -334,15 +332,13 @@ pub(crate) use self::testing::*;
 pub(crate) mod textcontrol;
 pub(crate) mod timeranges;
 pub(crate) mod touch;
-pub(crate) mod touchlist;
+pub(crate) use self::touch::*;
 pub(crate) mod trustedtypes;
 pub(crate) use self::trustedtypes::*;
 pub(crate) mod url;
 pub(crate) use self::url::*;
 pub(crate) mod useractivation;
 pub(crate) mod userscripts;
-pub(crate) mod validation;
-pub(crate) mod validitystate;
 pub(crate) mod values;
 pub(crate) mod visualviewport;
 pub(crate) mod wakelock;
