+++ components/script/dom/storage/storagemanager.rs
